Split store/data out into a store mixin
The store mixin shows how a mixin may be able to affect changes in behavior to the grid. Work needs to be done to use externalState (possibly) on rows and their subnodes. We also need to figure out a way for a subnode (body in this case) to communicate with the parent grid what new properties it’s adding, how to detect changes, and how to pass that data down through the hierarchy.